STREL-116 fix indel MQ0

+++ src/c++/lib/blt_common/MapqTracker.hh
@@ -67,7 +67,7 @@ struct MapqTracker
     getZeroFrac() const
     {
         if (count==0) return 0;
-        return (zeroCount/count);
+        return (zeroCount/static_cast<double>(count));
     }
 
     unsigned count = 0;
